This is a interview test assignment.
In this some task is given related to DOM manipulation, using some mathematical formulas. So first we need to create circles where the user click on screen. The clicked point will be the center. After that number of circles are no more than 2. After that both circles have random radius. And at last, check if the circles are intersecting or not.